Refactor tests to use jest
