+++ .vscode/launch.json
@@ -3,14 +3,10 @@
         {
             "type": "node",
             "request": "launch",
-            "name": "Mocha Current File",
-            "program": "${workspaceFolder}/node_modules/mocha/bin/_mocha",
+            "name": "Jest Current File",
+            "program": "${workspaceFolder}/node_modules/jest/bin/jest.js",
             "args": [
-              "--no-timeouts",
-              "--colors",
               "${file}",
-              "--require",
-              "ts-node/register"
             ],
             "sourceMaps": true,
             "preLaunchTask": "tsc",
