+++ package.json
@@ -5,8 +5,7 @@
   "main": "index.js",
   "types": "index.d.ts",
   "scripts": {
-    "test": "mocha -r ts-node/register tests/**/*.test.ts",
-    "coverage": "nyc -r lcov -e .ts -x \"*.test.ts\" npm run test"
+    "test": "jest"
   },
   "keywords": [],
   "author": "Dylan Ratcliffe",
@@ -22,16 +21,24 @@
     "ws": "^8.8.1"
   },
   "devDependencies": {
+    "@babel/core": "^7.19.0",
+    "@babel/preset-env": "^7.19.0",
+    "@babel/preset-typescript": "^7.18.6",
     "@types/chai": "^4.2.18",
     "@types/google-protobuf": "^3.7.4",
     "@types/mocha": "^8.2.2",
     "@types/sha1": "^1.1.2",
     "@types/uuid": "^8.3.3",
     "@types/ws": "^8.5.3",
+    "babel-jest": "^29.0.3",
     "chai": "^4.3.4",
     "create-index": "^2.6.0",
+    "jest": "^29.0.3",
+    "jest-environment-jsdom": "^29.0.3",
+    "jest-websocket-mock": "^2.4.0",
     "mocha": "^10.0.0",
     "nyc": "^15.1.0",
+    "ts-jest": "^29.0.0",
     "ts-node": "^10.0.0",
     "typescript": "^4.7.4"
   },
