+++ tests/unit/gateway.test.ts
@@ -1,0 +1,134 @@
+import WS from "jest-websocket-mock";
+import { describe, expect, it, beforeAll } from '@jest/globals';
+import { GatewaySession } from '../../gateway'
+import { Util } from '../../'
+import * as data from './sampledata'
+
+// create a WS instance
+const TestServerAddress = 'ws://localhost:31274'
+const server = new WS(TestServerAddress, {
+    jsonProtocol: false,
+});
+  
+describe('GatewaySession', () => {
+    describe('connecting', () => {
+        it('connects successfully', async () => {
+            var gs = new GatewaySession(TestServerAddress);
+            
+            await gs.ready
+            
+            expect(gs.state()).toBe(WebSocket.OPEN);
+            
+            gs.close();
+        })
+
+        it('sends requests', async () => {
+            var gs = new GatewaySession(TestServerAddress);
+            
+            await gs.ready
+
+            gs.sendRequest(data.gatewayRequest.itemRequest);
+
+            expect(await server.nextMessage).toEqual(data.gatewayRequest.itemRequest.serializeBinary());
+        })
+    })
+
+    describe('Processing inbound messages', () => {
+        var gs = new GatewaySession(TestServerAddress);
+        
+        beforeAll(async () => {
+            await gs.ready
+        });
+
+        describe('Error', () => {
+            it('should call the callback', (done) => {
+                var response = Util.newGatewayResponse("some error")
+
+                // Register the callbacks
+                gs.addEventListener(GatewaySession.ErrorEvent, (event) => {
+                    expect(event.detail).toEqual("some error")
+                    done();
+                }, { once: true })
+
+                server.send(response.serializeBinary().buffer);
+            })
+        })
+        describe('NewItem', () => {
+            it('should call the callback', (done) => {
+                var response = Util.newGatewayResponse(data.itemData.dylan)
+
+                // Register the callbacks
+                gs.addEventListener(GatewaySession.NewItemEvent, (event) => {
+                    expect(event.detail.getType()).toEqual(data.item.dylan.getType())
+                    done();
+                }, { once: true })
+
+                server.send(response.serializeBinary().buffer);
+            })
+        })
+        describe('NewEdge', () => {
+            it('should call the callback', (done) => {
+                var response = Util.newGatewayResponse(data.edgeData.basic)
+
+                // Register the callbacks
+                gs.addEventListener(GatewaySession.NewEdgeEvent, (event) => {
+                    expect(event.detail.getFrom()).toEqual(data.edge.basic.getFrom())
+                    expect(event.detail.getTo()).toEqual(data.edge.basic.getTo())
+                    done();
+                }, { once: true })
+
+                server.send(response.serializeBinary().buffer);
+            })
+        })
+        describe('NewItemRequestError', () => {
+            it('should call the callback', (done) => {
+                var response = Util.newGatewayResponse(data.errorData.NOCONTEXT)
+
+                // Register the callbacks
+                gs.addEventListener(GatewaySession.NewItemRequestErrorEvent, (event) => {
+                    expect(event.detail.getContext()).toEqual(data.error.NOCONTEXT.getContext());
+                    expect(event.detail.getErrortype()).toEqual(data.error.NOCONTEXT.getErrortype());
+                    done();
+                }, { once: true })
+
+                server.send(response.serializeBinary().buffer);
+            })
+        })
+        describe('Status', () => {
+            it('should call the callback', (done) => {
+                var response = Util.newGatewayResponse(data.gatewayStatusData.working)
+
+                // Register the callbacks
+                gs.addEventListener(GatewaySession.StatusEvent, (event) => {
+                    expect(event.detail.getPostprocessingcomplete()).toEqual(data.gatewayStatus.working.getPostprocessingcomplete());
+                    expect(event.detail.getSummary()?.toObject()).toEqual(data.gatewayStatus.working.getSummary()?.toObject());
+                    done();
+                }, {once: true})
+
+                server.send(response.serializeBinary().buffer);
+            })
+            
+            it('should update the status', (done) => {
+                var working = Util.newGatewayResponse(data.gatewayStatusData.working)
+                var doneResponse = Util.newGatewayResponse(data.gatewayStatusData.done)
+
+                // Register the callbacks
+                gs.addEventListener(GatewaySession.StatusEvent, () => {
+                    expect(gs.status?.postprocessingcomplete).toEqual(data.gatewayStatus.working.getPostprocessingcomplete());
+                    expect(gs.status?.summary).toEqual(data.gatewayStatus.working.getSummary()?.toObject());
+
+                    // Add the second test
+                    gs.addEventListener(GatewaySession.StatusEvent, () => {
+                        expect(gs.status?.postprocessingcomplete).toEqual(data.gatewayStatus.done.getPostprocessingcomplete());
+                        expect(gs.status?.summary).toEqual(data.gatewayStatus.done.getSummary()?.toObject());
+                        done();
+                    }, { once: true })
+
+                    server.send(doneResponse.serializeBinary().buffer);
+                }, { once: true })
+
+                server.send(working.serializeBinary().buffer);
+            })
+        })
+    })
+})
