+++ tests/unit/sampledata.ts
@@ -1,6 +1,6 @@
-import { Util, RequestMethod } from '../index'
-import { ResponderState } from "../responses_pb";
-import { ItemRequestError } from "../responses_pb";
+import { Util, RequestMethod } from '../../index'
+import { ResponderState } from "../../responses_pb";
+import { ItemRequestError } from "../../responses_pb";
 
 export const errorData = {
     NOTFOUND: {
